chore: refactor and remove all depedencies
