+++ composer.json
@@ -12,8 +12,7 @@
     "source": "https://github.com/itsnubix/nowcal"
   },
   "require": {
-    "php": "^8.1",
-    "ramsey/uuid": ">3.8"
+    "php": "^7.0"
   },
   "require-dev": {
     "friendsofphp/php-cs-fixer": "^3.0",
